+++ packages/date-picker/src/components/day-names.tsx
@@ -1,6 +1,9 @@
 import { useContext } from 'preact/compat'
 import { AppContext } from '../utils/stateful/app-context'
-import { getOneLetterDayNames } from '../../../../shared/utils/stateless/time/date-time-localization/date-time-localization'
+import {
+  getDayNamesShort,
+  getOneLetterDayNames,
+} from '../../../../shared/utils/stateless/time/date-time-localization/date-time-localization'
 import { toJSDate } from '../../../../shared/utils/stateless/time/format-conversion/format-conversion'
 
 export default function DayNames() {
@@ -8,7 +11,12 @@ export default function DayNames() {
   const aWeek = $app.timeUnitsImpl.getWeekFor(
     toJSDate($app.datePickerState.datePickerDate.value)
   )
-  const dayNames = getOneLetterDayNames(aWeek, $app.config.locale)
+  let dayNames: string[] = []
+  if (['zh-cn'].includes($app.config.locale.toLowerCase())) {
+    dayNames = getDayNamesShort(aWeek, $app.config.locale)
+  } else {
+    dayNames = getOneLetterDayNames(aWeek, $app.config.locale)
+  }
 
   return (
     <div class="sx__date-picker__day-names">