+++ packages/date-picker/src/utils/stateless/testing/create-app-singleton.ts
@@ -4,3 +4,8 @@ export const appSingletonWithGerman = () =>
   createAppSingleton({
     locale: 'de-DE',
   })
+
+export const appSingletonWithLocale = (locale: string) =>
+  createAppSingleton({
+    locale,
+  })