+++ packages/translations/src/locales/sv-SE/date-picker.ts
@@ -1,0 +1,8 @@
+import { DatePickerTranslations } from '../../types/date-picker.translations'
+
+export const datePickerSvSE: DatePickerTranslations = {
+  Date: 'Datum',
+  'MM/DD/YYYY': 'ÅÅÅÅ-MM-DD',
+  'Next month': 'Nästa månad',
+  'Previous month': 'Föregående månad',
+}