+++ packages/translations/src/locales/sv-SE/index.ts
@@ -1,0 +1,6 @@
+import { datePickerSvSE } from './date-picker'
+import { Language } from '../../types/language.translations'
+
+export const svSE: Language = {
+  ...datePickerSvSE,
+}