+++ shared/utils/stateless/time/format-conversion/date-format/date-format-localized-rules.ts
@@ -17,6 +17,10 @@ const formatRules = {
     delimiter: DateFormatDelimiter.SLASH,
     order: DateFormatOrder.DMY,
   },
+  slashYMD: {
+    delimiter: DateFormatDelimiter.SLASH,
+    order: DateFormatOrder.YMD,
+  },
   periodDMY: {
     delimiter: DateFormatDelimiter.PERIOD,
     order: DateFormatOrder.DMY,
@@ -30,6 +34,7 @@ const formatRules = {
 export const dateFormatLocalizedRules = new Map<string, DateFormatRule>([
   ['en-US', formatRules.slashMDY],
   ['en-GB', formatRules.slashDMY],
+  ['zh-CN', formatRules.slashYMD],
   ['de-DE', formatRules.periodDMY],
   ['sv-SE', formatRules.dashYMD],
 ])