+++ packages/theme-default/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.1.0-alpha.3 (2023-08-27)
+
+
+### Bug Fixes
+
+* svg compilation ([#17](https://github.com/schedule-x/schedule-x/issues/17)) ([493e9ba](https://github.com/schedule-x/schedule-x/commit/493e9ba6bf05834902b1dade8baeda60774268c8))
+
+
+### Features
+
+* add placement config ([#9](https://github.com/schedule-x/schedule-x/issues/9)) ([3344d69](https://github.com/schedule-x/schedule-x/commit/3344d69c042ba09e875aa80c2497c2f6bf336bac))
+* allow empty string value ([#14](https://github.com/schedule-x/schedule-x/issues/14)) ([94aa1b9](https://github.com/schedule-x/schedule-x/commit/94aa1b92375f7b8409d7008fbed599bc02fc46ae))
+* create month view for date picker ([#1](https://github.com/schedule-x/schedule-x/issues/1)) ([52d05c7](https://github.com/schedule-x/schedule-x/commit/52d05c788fbd215ac9f416f1051ad9cf0e172e1e))
+* dark mode ([#19](https://github.com/schedule-x/schedule-x/issues/19)) ([f3b6a4a](https://github.com/schedule-x/schedule-x/commit/f3b6a4ac467ce4af0498019090c79e75763beb90))
+* date picker min max ([#5](https://github.com/schedule-x/schedule-x/issues/5)) ([b92aafe](https://github.com/schedule-x/schedule-x/commit/b92aafec67608356384f047383d6f38d95323902))
+* implement date picker years view ([#3](https://github.com/schedule-x/schedule-x/issues/3)) ([8ff0489](https://github.com/schedule-x/schedule-x/commit/8ff04898ffb6ccdc2a66a27532f25a51d865ed39))
+* make it possible to select a date ([#4](https://github.com/schedule-x/schedule-x/issues/4)) ([b5fb130](https://github.com/schedule-x/schedule-x/commit/b5fb13072da25b4a13c0236e0095b473ebb5f8fb))
+* popup position ([#12](https://github.com/schedule-x/schedule-x/issues/12)) ([fcf93fc](https://github.com/schedule-x/schedule-x/commit/fcf93fcbe147b1c162322f32b7cfced2bb484ae6))
+* style leading and trailing dates differently ([#11](https://github.com/schedule-x/schedule-x/issues/11)) ([a2ef08a](https://github.com/schedule-x/schedule-x/commit/a2ef08ab35a60f9eb4b244fc1a681a33a54d4bdf))
+
+
+
+
+
 # 0.1.0-alpha.2 (2023-08-18)
 
 ### Features